+++ src/apis/http/governance.js
@@ -2,22 +2,6 @@ import { APIS } from '@/constants';
 import axios from 'axios';
 
 const governance = {
-  requestList(status) {
-    const params = { proposalStatus: status };
-    return axios({
-      url: `/proposals`,
-      method: 'get',
-      baseURL: `${APIS.HTTP.GOVERNANCE}`,
-      params,
-    });
-  },
-  requestDetail(id) {
-    return axios({
-      url: `/proposals/${id}`,
-      method: 'get',
-      baseURL: `${APIS.HTTP.GOVERNANCE}`,
-    });
-  },
   requestTally(id) {
     return axios({
       url: `/proposals/${id}/tally`,