+++ src/components/layout/TheFooter.vue
@@ -18,15 +18,25 @@
       >
         <v-icon dark>{{ mdiArrowUpBold }}</v-icon>
       </v-btn>
-      <v-row class="my-0 py-0 py-sm-2 overline rm__text-transform">
-        <v-col cols="12" sm="2" class="my-0 py-0">
-          <span
-            class="d-flex justify-center justify-sm-start"
-            v-text="appVersion"
-          />
+      <v-row class="d-flex align-center rm__text-transform">
+        <v-col cols="12" md="4" order="1" order-md="3" class="pb-1 py-md-5">
+          <div class="text-center text-md-right text-caption">
+            <span class="pr-1 font-weight-bold" v-text="chainNetwork" />
+            <span v-text="chainVersion" />
+            <v-icon right>
+              {{ connectionIcon }}
+            </v-icon>
+          </div>
         </v-col>
-        <v-col cols="12" sm="8" class="my-0 py-0">
-          <span class="font-italic d-flex justify-center" v-html="copy" />
+        <v-col cols="12" md="4" order="2" class="py-0 py-md-5">
+          <div class="text-center text-overline font-weight-bold">
+            <span class="font-italic" v-html="copy" />
+          </div>
+        </v-col>
+        <v-col cols="12" md="4" order="3" order-md="1" class="pt-1 py-md-5">
+          <div class="text-center text-md-left text-caption">
+            <span v-text="appVersion" />
+          </div>
         </v-col>
       </v-row>
     </v-footer>
@@ -35,7 +45,8 @@
 
 <script>
 import { CONFIG } from '@/constants';
-import { mdiArrowUpBold } from '@mdi/js';
+import { mapGetters } from 'vuex';
+import { mdiArrowUpBold, mdiLanConnect, mdiLanDisconnect } from '@mdi/js';
 
 export default {
   name: 'TheFooter',
@@ -46,8 +57,24 @@ export default {
     };
   },
   computed: {
+    ...mapGetters('application', ['info']),
     appVersion() {
       return `v${process.env.APP_VERSION}`;
+    },
+    chainNetwork() {
+      return this.info && this.info.node_info && this.info.node_info.network
+        ? this.info.node_info.network
+        : '';
+    },
+    chainVersion() {
+      return this.info &&
+        this.info.application_version &&
+        this.info.application_version.version
+        ? `(${this.info.application_version.version})`
+        : '';
+    },
+    connectionIcon() {
+      return this.info ? mdiLanConnect : mdiLanDisconnect;
     },
     copy() {
       const year = new Date().getFullYear();