+++ src/components/layout/TheNavBar.vue
@@ -5,30 +5,20 @@
     </v-navigation-drawer>
     <v-app-bar app color="primary" dark flat>
       <v-row class="d-flex align-center">
-        <v-col cols="2" md="4" class="d-flex justify-start">
+        <v-col cols="2" md="5" class="d-flex justify-start">
           <v-app-bar-nav-icon @click.stop="drawer = !drawer">
             <slot>
               <v-icon size="35">{{ mdiMenu }}</v-icon>
             </slot>
           </v-app-bar-nav-icon>
         </v-col>
-        <v-col
-          cols="2"
-          md="4"
-          class="d-flex justify-space-between justify-md-center"
-        >
+        <v-col cols="3" md="2" class="d-flex justify-center">
           <router-link :to="{ name: ROUTES.NAME.HOME }">
             <v-img :src="logoSrc" alt="Logo" width="50" />
           </router-link>
         </v-col>
-        <v-col cols="8" md="4" class="d-flex justify-end">
-          <div class="caption text-right">
-            <span class="pr-1 font-weight-bold" v-text="chainNetwork" />
-            <span v-text="chainVersion" />
-          </div>
-          <v-icon right>
-            {{ connectionIcon }}
-          </v-icon>
+        <v-col cols="7" md="5" class="d-flex justify-end">
+          <TheConnectWalletButton />
         </v-col>
       </v-row>
     </v-app-bar>
@@ -36,15 +26,16 @@
 </template>
 
 <script>
-import TheDrawer from './drawer/TheDrawer';
+import TheConnectWalletButton from './nav-bar/TheConnectWalletButton.vue';
+import TheDrawer from './nav-bar/drawer/TheDrawer';
 
 import { ROUTES } from '@/constants';
-import { mapGetters } from 'vuex';
-import { mdiLanConnect, mdiLanDisconnect, mdiMenu } from '@mdi/js';
+import { mdiMenu } from '@mdi/js';
 
 export default {
   name: 'TheNavBar',
   components: {
+    TheConnectWalletButton,
     TheDrawer,
   },
   data() {
@@ -55,24 +46,8 @@ export default {
     };
   },
   computed: {
-    ...mapGetters('application', ['info']),
     logoSrc() {
       return require('@/assets/img/logo-white.png');
-    },
-    connectionIcon() {
-      return this.info ? mdiLanConnect : mdiLanDisconnect;
-    },
-    chainNetwork() {
-      return this.info && this.info.node_info && this.info.node_info.network
-        ? this.info.node_info.network
-        : '';
-    },
-    chainVersion() {
-      return this.info &&
-        this.info.application_version &&
-        this.info.application_version.version
-        ? `(${this.info.application_version.version})`
-        : '';
     },
   },
 };